GEN_AI/RESUME_SCREENER_TASK_3/requirements.txt
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-huggingface

huggingface_hub

langsmith

python-dotenv
pydantic